Introduce the ANN HOW getter
In the context of the MOP, the type object used to retrieve a HOW's
annotations attribute via yield_annotations is dynamic, making ANN more
or less pointless. On top of that, ANN is geared more towards when the
type object is known; Data::Record wants to depend on annotations in the
context of an archetypes metamethod, where a type object isn't always
known, so we need to be able to know the annotations for a HOW given the
HOW itself too. Such an ANN getter on a HOW can back yield_annotations
by default, which is to be its more dynamic sibling from now on.